backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
geoalchemy2
shapely
pydantic-settings
email-validator
pytest
httpx
pytest-asyncio
